+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.1.1]
+
+- Added Atom One themes in the extension pack
+- Added theme selection commands
+- Commands are now on workspace level instead of global level
+- Setup of the workspace settings is now executed with a command instead when the extension is loaded, due the change above
+
 ## [0.0.37]
 
 - Added setup of default formatter